Move remove_file to utils
