+++ resources/signals.py
@@ -1,11 +1,9 @@
-import logging
 import os
 import shutil
 
 from django.conf import settings
 from django.db.models.signals import post_delete, pre_save
 from django.dispatch import receiver
-from sentry_sdk import capture_exception
 
 from resources.models import (
     get_harbor_media_folder,
@@ -15,18 +13,7 @@ from resources.models import (
     WinterStorageArea,
     WinterStorageAreaMap,
 )
-
-logger = logging.getLogger(__name__)
-
-
-def remove_file(instance, field_name):
-    file = getattr(instance, field_name, None)
-    if file:
-        try:
-            os.unlink(file.path)
-        except FileNotFoundError as e:
-            logger.error(e, exc_info=True)
-            capture_exception(e)
+from utils.files import remove_file
 
 
 @receiver(post_delete, sender=HarborMap)
